refactor: update import for OKTO_REMOTE_CONFIG to use localConfig.js
